+++ src/graphic/rendertarget.h
@@ -62,18 +62,18 @@ public:
 	void fill_rect(const Rectf&, const RGBAColor&, BlendMode blend_mode = BlendMode::Copy);
 	void brighten_rect(const Rectf&, int32_t factor);
 
-	void blit(const Vector2f& dst,
+	void blit(const Vector2i& dst,
 	          const Image* image,
 	          BlendMode blend_mode = BlendMode::UseAlpha,
 	          UI::Align = UI::Align::kLeft);
 
 	// Like blit. See MonochromeBlitProgram for details.
-	void blit_monochrome(const Vector2f& dst,
+	void blit_monochrome(const Vector2i& dst,
 	                     const Image* image,
 	                     const RGBAColor& blend_mode,
 	                     UI::Align = UI::Align::kLeft);
 
-	void blitrect(const Vector2f& dst,
+	void blitrect(const Vector2i& dst,
 	              const Image* image,
 	              const Recti& src,
 	              BlendMode blend_mode = BlendMode::UseAlpha);
