+++ src/wui/interactive_base.cc
@@ -358,7 +358,7 @@ void InteractiveBase::draw_overlay(RenderTarget& dst) {
 		if (is_game) {
 			const std::string gametime(gametimestring(egbase().get_gametime(), true));
 			const std::string gametime_text = as_condensed(gametime);
-			dst.blit(Vector2f(5, 5), UI::g_fh1->render(gametime_text), BlendMode::UseAlpha,
+			dst.blit(Vector2i(5, 5), UI::g_fh1->render(gametime_text), BlendMode::UseAlpha,
 			         UI::Align::kLeft);
 
 			static boost::format node_format("(%i, %i)");
@@ -371,7 +371,7 @@ void InteractiveBase::draw_overlay(RenderTarget& dst) {
 
 		const Image* rendered_text = UI::g_fh1->render(node_text);
 
-		dst.blit(Vector2f(get_w() - 5, get_h() - rendered_text->height() - 5), rendered_text,
+		dst.blit(Vector2i(get_w() - 5, get_h() - rendered_text->height() - 5), rendered_text,
 		         BlendMode::UseAlpha, UI::Align::kRight);
 	}
 
@@ -380,7 +380,7 @@ void InteractiveBase::draw_overlay(RenderTarget& dst) {
 		static boost::format fps_format("%5.1f fps (avg: %5.1f fps)");
 		const Image* rendered_text = UI::g_fh1->render(as_condensed(
 		   (fps_format % (1000.0 / frametime_) % (1000.0 / (avg_usframetime_ / 1000))).str()));
-		dst.blit(Vector2f((get_w() - rendered_text->width()) / 2, 5), rendered_text,
+		dst.blit(Vector2i((get_w() - rendered_text->width()) / 2, 5), rendered_text,
 		         BlendMode::UseAlpha, UI::Align::kLeft);
 	}
 }
