+++ src/wui/interactive_gamebase.cc
@@ -126,7 +126,7 @@ void InteractiveGameBase::draw_overlay(RenderTarget& dst) {
 		}
 
 		if (!game_speed.empty()) {
-			dst.blit(Vector2f(get_w() - 5, 5), UI::g_fh1->render(game_speed), BlendMode::UseAlpha,
+			dst.blit(Vector2i(get_w() - 5, 5), UI::g_fh1->render(game_speed), BlendMode::UseAlpha,
 			         UI::Align::kRight);
 		}
 	}
